Add authorization gates for features

+++ app/Http/Controllers/FeatureController.php
@@ -6,12 +6,16 @@ use App\Http\Requests\Feature\StoreFeatureRequest;
 use App\Http\Requests\Feature\UpdateFeatureRequest;
 use App\Models\Feature;
 use App\Models\Project;
+use Gate;
 use Illuminate\Http\Request;
 
 class FeatureController extends Controller
 {
 
-    public function list(Request $request, $projectId){
+    public function list(Request $request, $projectId)
+    {
+        Gate::authorize('feature.list', $projectId);
+
         $eagerload = $request->has('eagerload');
         $query = Feature::query()->where('project_id', $projectId);
 
@@ -26,6 +30,8 @@ class FeatureController extends Controller
      */
     public function store(StoreFeatureRequest $request, $projectId)
     {
+        Gate::authorize('feature.create', $projectId);
+        
         $project = Project::findOrFail($projectId);
         $feature = $project->features()->create($request->only('name', 'description', 'links'));
         return response()->json($feature, 201)->header('Content-Type','application/json');
@@ -36,6 +42,8 @@ class FeatureController extends Controller
      */
     public function show(Request $request, $id)
     {
+        Gate::authorize('feature.view', $id);
+
         $eagerload = $request->has('eagerload');
         $query = Feature::query();
 
@@ -51,6 +59,8 @@ class FeatureController extends Controller
      */
     public function update(UpdateFeatureRequest $request, $id)
     {
+        Gate::authorize('feature.update/delete', $id);
+
         $feature = Feature::where('id', $id)->firstOrFail();
         $feature->update($request->only('name', 'description', 'links'));
         return response()->json($feature, 200)->header('Content-Type','application/json');
@@ -61,6 +71,8 @@ class FeatureController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('feature.update/delete', $id);
+
         $feature = Feature::where('id',$id)->firstOrFail();
         $feature->delete();
         return response()->json([], 204)->header('Content-Type','application/json');
